feat: :art: send notification
